pyproject.toml
[project]
name = "webqa-plus"
version = "1.0.0"
description = "Best-of-all-worlds autonomous AI web QA tester with Playwright MCP, LangGraph, and Rich visual reporting"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    {name = "WebQA-Plus Team"}
]
keywords = ["testing", "qa", "playwright", "ai", "automation", "langgraph"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "playwright>=1.41.0",
    "playwright-stealth>=1.0.0",
    "langgraph>=0.0.50",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.0.20",
    "weasyprint>=60.0",
    "jinja2>=3.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "pillow>=10.0.0",
    "aiohttp>=3.9.0",
    "tenacity>=8.2.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "requests>=2.31.0",
    "mcp>=1.0.0",
    "playwright-mcp>=0.0.1",
    "webqa-agent @ git+https://github.com/MigoXLab/webqa-agent.git@main",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
]

[project.scripts]
webqa-plus = "webqa_plus.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/webqa-plus"
Repository = "https://github.com/yourusername/webqa-plus"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["src/webqa_plus"]
